fix: 타입 수정

+++ src/main/java/pretzel/dreamketcherbe/domain/comment/dto/RecommentResDto.java
@@ -9,7 +9,7 @@ public record RecommentResDto(
     String ninkname,
     String content,
     Long parentCommentId,
-    Long commentOrder
+    int commentOrder
 ) {
 
     public static RecommentResDto of(Recomment recomment) {
